ai_trading_system/requirements.txt
# AI Trading System Requirements

# Core Dependencies
numpy>=1.21.0
pandas>=1.5.0
torch>=2.0.0
scikit-learn>=1.0.0

# API & Networking
requests>=2.28.0
aiohttp>=3.8.0
websockets>=10.0

# Technical Analysis
ta>=0.10.0
ta-lib>=0.4.25

# Machine Learning
optuna>=3.0.0
lightgbm>=3.3.0
xgboost>=1.7.0

# Database & Caching
redis>=4.5.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Monitoring & Logging
streamlit>=1.20.0
plotly>=5.13.0
mlflow>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
schedule>=1.1.0

# Development
pytest>=7.2.0
pytest-asyncio>=0.20.0
black>=23.0.0
flake8>=6.0.0